docs(regex.translation): expand/simplify regex translation feature docs (#379)
The **Regex Translation** Framework component had both too little and
too much information as it talked about internal workings - like 'on
start-up' - while failed to mention the capabilities of the translation
itself.

This PR tries to simplify and expand on certain topics from an user's
perspective to make the feature documentation more usable.