docs(flow.trace): correct flow trace search visualization (#399)
The feature documentation page explaining the flow trace capabilities
was incomplete on some parts (now showing/explaining DSAV, for example)
and too detailed on others (deep-dive into regular expressions).

This PR tries with `.gif` images to unburden the text and with Markdown
tables to more easily explain the possibilities for the flow trace
search.

Closes #397