chore(flow): centralize flow crud on single page (#406)
Previously, the flow creation, alerting and permissions were all split
into multiple pages, while they were talking about the same thing/on the
same Dashboard UI page. This was already improved somewhat by placing
the alerting and flow creation/editing on the same page, but the flow
permissions were still separated on a smaller page that provided less
added-value to be separated.

This PR centralizes further anything related to 'CRUD actions' on flows.
It also removes the need of 'textual breadcrumbs' in favor of an
illustration.