chore(seqctrl): streamline motivation & endpoints (#413)
Streamlines how the 'motivation' and 'endpoints' section in the other
Framework components are represented in the Sequence Controller
component.

Plus, removes the unnecessary Logic Apps setup, as this only shows what
HTTP webhook and calls look like, which is rather standard, is easy to
get 'deprecated' or 'changed' if the Logic App designer changes, and
does not add much value besides the existing 'pseudo-la' diagram.