chore(transco.v2): move matrix basic promote v2 to transco v2 page (#412)
The Matrix basic promote functionality was previously separated from the
main Transco v2 documentation page. This was previously done probably to
not overwhelm the single Transco v2 page (with Netlify). Now, with
Docusaurus, tabs can be used for parallel content - which is what this
is.

This PR moves the Matrix basic promote functionality to the Transco v2
page, so that everything Transco-related can be found on the same page.