chore(installation): use installation wizard for dashboard/framework (#417)
Since the installation of both the Framework and Dashboard spanned
several pages and YAML template files, it was both rather hard to
navigate through everything as well as maintaining the whole thing.

This PR radically changes this approach of different files to a single
installation page for the Framework and one for the Dashboard. Using a
visually attractive and clickable installation wizard and information
badges to show 'shared' sections between the installations, the entire
process is simplified quite a lot.

<img width="922" height="780" alt="image"
src="https://github.com/user-attachments/assets/19ef1a8d-fa81-4d1d-9b05-482a14a84166"
/>

---------

Co-authored-by: Pim Simons <32359437+pim-simons@users.noreply.github.com>