chore(sequencing): streamling seq ctrl & time seq w workflow runs diagrams (#422)
Streamline the diagrams, style and title for Azure Logic App workflow
runs in the **Sequence Controller** and **Time Sequencer**. This helps
with understanding the similarities between the two components, as well
as the possibilities. This was previously harder to grasp with only an
overview of a single Logic App pseudo definition.