fix(link): use correct links for framework docs (#426)
Due to late merges with the main branch and parallel PR's, some broken
links were only discovered when trying to release the new version from
the main branch -- instead of being halted at PR-time.

This PR fixes the broken links in the Framework docs related to the
XML/JSON Converter and the installation page.