chore(roles): provide a more in-depth role page / func (#431)
The 'role management' page was both rather limited in its explanation as
well as difficult to map to actual Dashboard actions.

This PR provides tables for flow permissions and user roles and map this
to specific Dashboard actions, possibly with some descriptive context.
Placing the flow permissions first also follows the 'main' reason of the
Dashboard, as well as the CRUD actions that are now split.