chore(flowtrace.la): use walkthrough for import flow traces via la (#448)
There are several steps that consumers need to follow for them to
successfully import their flow traces into the Dashboard. The
non-numbered sections were not that clear that this is part of a broader
walkthrough.

This PR separates the 'importing' from the 'usage' section, by making
sure that we provide a walkthrough for the 'importing' and leaving the
'usage' section directly visible.